Started working on netwocking packets, basic serialization and desirialization alhoritms implemented

+++ src/messages/MessageUtils.hpp
@@ -3,5 +3,5 @@
 #include <cstdint>
 #include <memory>
 
-int generate_unique_id();
+uint64_t generate_unique_id();
 int64_t generate_unique_timestamp();
